🚀 release: Update versions (#325)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## @tylerbu/cli@0.4.1

### Patch Changes

- Add missing template files to package
_[`#324`](https://github.com/tylerbutler/tools-monorepo/pull/324)
[`cc72e15`](https://github.com/tylerbutler/tools-monorepo/commit/cc72e156f12de4f502c05a57eeee44d05195ff17)
[@tylerbutler](https://github.com/tylerbutler)_

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>